+++ src/littletree/exceptions.py
@@ -27,13 +27,3 @@ class LoopError(TreeException):
 
     def __str__(self):
         return f"{self.node} can't become parent of {self.ancestor}"
-
-
-# TODO Currently unused. It was used in old Route-class
-class DifferentTreeError(TreeException):
-    def __init__(self, root1, root2):
-        self.root1 = root1
-        self.root2 = root2
-
-    def __str__(self):
-        return f"Nodes have different roots: {self.root1} vs {self.root2}"